moved socket handling to its own module
